Adds Abstract class for AASServerComponentFactory * Renames Interface
