+++ basyx.components/basyx.components.docker/basyx.components.AASServer/src/main/java/org/eclipse/basyx/components/aas/aascomponent/InMemoryAASServerComponentFactory.java
@@ -12,7 +12,6 @@ package org.eclipse.basyx.components.aas.aascomponent;
 import java.util.List;
 
 import org.eclipse.basyx.aas.aggregator.AASAggregatorFactory;
-import org.eclipse.basyx.aas.aggregator.api.IAASAggregator;
 import org.eclipse.basyx.aas.aggregator.api.IAASAggregatorFactory;
 import org.eclipse.basyx.aas.registration.api.IAASRegistry;
 import org.eclipse.basyx.aas.restapi.api.IAASAPIFactory;
@@ -30,10 +29,7 @@ import org.eclipse.basyx.submodel.restapi.vab.VABSubmodelAPIFactory;
  * @author fischer, fried
  *
  */
-public class InMemoryAASServerComponentFactory implements IAASServerComponentAggregatorFactory {
-
-	private List<IAASServerDecorator> aasServerDecorators;
-	private IAASRegistry aasServerRegistry;
+public class InMemoryAASServerComponentFactory extends AbstractAASServerComponentFactory {
 
 	public InMemoryAASServerComponentFactory(List<IAASServerDecorator> decorators, IAASRegistry aasServerRegistry) {
 		this.aasServerRegistry = aasServerRegistry;
@@ -44,68 +40,19 @@ public class InMemoryAASServerComponentFactory implements IAASServerComponentAgg
 		this.aasServerRegistry = aasServerRegistry;
 	}
 
-	@Override
-	public IAASAggregator create() {
-		ISubmodelAPIFactory submodelAPIFactory = createAndDecorateSubmodelAPIFactory();
-
-		ISubmodelAggregatorFactory submodelAggregatorFactory = createAndDecorateSubmodelAggregatorFactory(submodelAPIFactory);
-
-		IAASAPIFactory aasAPIFactory = createAndDecorateAASAPIFactory();
-
-		IAASAggregatorFactory aasAggregatorFactory = createAndDecorateAASAggregatorFactory(aasAPIFactory, submodelAggregatorFactory);
-
-		return aasAggregatorFactory.create();
-	}
-
-	private ISubmodelAPIFactory createAndDecorateSubmodelAPIFactory() {
-		ISubmodelAPIFactory submodelAPIFactory = createSubmodelAPIFactory();
-		for (IAASServerDecorator aasServerDecorator : aasServerDecorators) {
-			submodelAPIFactory = aasServerDecorator.decorateSubmodelAPIFactory(submodelAPIFactory);
-		}
-
-		return submodelAPIFactory;
-	}
-
-	private ISubmodelAggregatorFactory createAndDecorateSubmodelAggregatorFactory(ISubmodelAPIFactory submodelAPIFactory) {
-		ISubmodelAggregatorFactory submodelAggregatorFactory = createSubmodelAggregatorFactory(submodelAPIFactory);
-		for (IAASServerDecorator aasServerDecorator : aasServerDecorators) {
-			submodelAggregatorFactory = aasServerDecorator.decorateSubmodelAggregatorFactory(submodelAggregatorFactory);
-		}
-
-		return submodelAggregatorFactory;
-	}
-
-	private IAASAPIFactory createAndDecorateAASAPIFactory() {
-		IAASAPIFactory aasAPIFactory = createAASAPIFactory();
-		for (IAASServerDecorator aasServerDecorator : aasServerDecorators) {
-			aasAPIFactory = aasServerDecorator.decorateAASAPIFactory(aasAPIFactory);
-		}
-
-		return aasAPIFactory;
-	}
-
-	private IAASAggregatorFactory createAndDecorateAASAggregatorFactory(IAASAPIFactory aasAPIFactory, ISubmodelAggregatorFactory submodelAggregatorFactory) {
-		IAASAggregatorFactory aasAggregatorFactory = createAASAggregatorFactory(aasAPIFactory, submodelAggregatorFactory);
-		for (IAASServerDecorator aasServerDecorator : aasServerDecorators) {
-			aasAggregatorFactory = aasServerDecorator.decorateAASAggregatorFactory(aasAggregatorFactory);
-		}
-
-		return aasAggregatorFactory;
-	}
-
-	private ISubmodelAPIFactory createSubmodelAPIFactory() {
+	protected ISubmodelAPIFactory createSubmodelAPIFactory() {
 		return new VABSubmodelAPIFactory();
 	}
 
-	private ISubmodelAggregatorFactory createSubmodelAggregatorFactory(ISubmodelAPIFactory submodelAPIFactory) {
+	protected ISubmodelAggregatorFactory createSubmodelAggregatorFactory(ISubmodelAPIFactory submodelAPIFactory) {
 		return new SubmodelAggregatorFactory(submodelAPIFactory);
 	}
 
-	private IAASAPIFactory createAASAPIFactory() {
+	protected IAASAPIFactory createAASAPIFactory() {
 		return new VABAASAPIFactory();
 	}
 
-	private IAASAggregatorFactory createAASAggregatorFactory(IAASAPIFactory aasAPIFactory, ISubmodelAggregatorFactory submodelAggregatorFactory) {
+	protected IAASAggregatorFactory createAASAggregatorFactory(IAASAPIFactory aasAPIFactory, ISubmodelAggregatorFactory submodelAggregatorFactory) {
 		return new AASAggregatorFactory(aasAPIFactory, submodelAggregatorFactory, aasServerRegistry);
 	}
 }
