+++ basyx.components/basyx.components.docker/basyx.components.AASServer/src/main/java/org/eclipse/basyx/components/aas/aascomponent/MongoDBAASServerComponentFactory.java
@@ -11,7 +11,6 @@ package org.eclipse.basyx.components.aas.aascomponent;
 
 import java.util.List;
 
-import org.eclipse.basyx.aas.aggregator.api.IAASAggregator;
 import org.eclipse.basyx.aas.aggregator.api.IAASAggregatorFactory;
 import org.eclipse.basyx.aas.registration.api.IAASRegistry;
 import org.eclipse.basyx.aas.restapi.api.IAASAPIFactory;
@@ -31,24 +30,9 @@ import org.eclipse.basyx.submodel.restapi.api.ISubmodelAPIFactory;
  * @author fischer, fried
  *
  */
-public class MongoDBAASServerComponentFactory implements IAASServerComponentAggregatorFactory {
+public class MongoDBAASServerComponentFactory extends AbstractAASServerComponentFactory {
 
-	private List<IAASServerDecorator> aasServerDecorators;
-	private IAASRegistry aasServerRegistry;
 	private BaSyxMongoDBConfiguration mongoDBConfig;
-
-	@Override
-	public IAASAggregator create() {
-		ISubmodelAPIFactory submodelAPIFactory = createAndDecorateSubmodelAPIFactory();
-
-		ISubmodelAggregatorFactory submodelAggregatorFactory = createAndDecorateSubmodelAggregatorFactory(submodelAPIFactory);
-
-		IAASAPIFactory aasAPIFactory = createAndDecorateAASAPIFactory();
-
-		IAASAggregatorFactory aasAggregatorFactory = createAndDecorateAASAggregatorFactory(aasAPIFactory, submodelAggregatorFactory);
-
-		return aasAggregatorFactory.create();
-	}
 
 	public MongoDBAASServerComponentFactory(BaSyxMongoDBConfiguration config, List<IAASServerDecorator> decorators, IAASRegistry aasServerRegistry) {
 		this.mongoDBConfig = config;
@@ -61,55 +45,19 @@ public class MongoDBAASServerComponentFactory implements IAASServerComponentAggr
 		this.aasServerRegistry = aasServerRegistry;
 	}
 
-	private ISubmodelAPIFactory createAndDecorateSubmodelAPIFactory() {
-		ISubmodelAPIFactory submodelAPIFactory = createSubmodelAPIFactory();
-		for (IAASServerDecorator aasServerDecorator : aasServerDecorators) {
-			submodelAPIFactory = aasServerDecorator.decorateSubmodelAPIFactory(submodelAPIFactory);
-		}
-
-		return submodelAPIFactory;
-	}
-
-	private ISubmodelAggregatorFactory createAndDecorateSubmodelAggregatorFactory(ISubmodelAPIFactory submodelAPIFactory) {
-		ISubmodelAggregatorFactory submodelAggregatorFactory = createSubmodelAggregatorFactory(submodelAPIFactory);
-		for (IAASServerDecorator aasServerDecorator : aasServerDecorators) {
-			submodelAggregatorFactory = aasServerDecorator.decorateSubmodelAggregatorFactory(submodelAggregatorFactory);
-		}
-
-		return submodelAggregatorFactory;
-	}
-
-	private IAASAPIFactory createAndDecorateAASAPIFactory() {
-		IAASAPIFactory aasAPIFactory = createAASAPIFactory();
-		for (IAASServerDecorator aasServerDecorator : aasServerDecorators) {
-			aasAPIFactory = aasServerDecorator.decorateAASAPIFactory(aasAPIFactory);
-		}
-
-		return aasAPIFactory;
-	}
-
-	private IAASAggregatorFactory createAndDecorateAASAggregatorFactory(IAASAPIFactory aasAPIFactory, ISubmodelAggregatorFactory submodelAggregatorFactory) {
-		IAASAggregatorFactory aasAggregatorFactory = createAASAggregatorFactory(aasAPIFactory, submodelAggregatorFactory);
-		for (IAASServerDecorator aasServerDecorator : aasServerDecorators) {
-			aasAggregatorFactory = aasServerDecorator.decorateAASAggregatorFactory(aasAggregatorFactory);
-		}
-
-		return aasAggregatorFactory;
-	}
-
-	private ISubmodelAPIFactory createSubmodelAPIFactory() {
+	protected ISubmodelAPIFactory createSubmodelAPIFactory() {
 		return new MongoDBSubmodelAPIFactory(mongoDBConfig);
 	}
 
-	private ISubmodelAggregatorFactory createSubmodelAggregatorFactory(ISubmodelAPIFactory submodelAPIFactory) {
+	protected ISubmodelAggregatorFactory createSubmodelAggregatorFactory(ISubmodelAPIFactory submodelAPIFactory) {
 		return new SubmodelAggregatorFactory(submodelAPIFactory);
 	}
 
-	private IAASAPIFactory createAASAPIFactory() {
+	protected IAASAPIFactory createAASAPIFactory() {
 		return new MongoDBAASAPIFactory(mongoDBConfig);
 	}
 
-	private IAASAggregatorFactory createAASAggregatorFactory(IAASAPIFactory aasAPIFactory, ISubmodelAggregatorFactory submodelAggregatorFactory) {
+	protected IAASAggregatorFactory createAASAggregatorFactory(IAASAPIFactory aasAPIFactory, ISubmodelAggregatorFactory submodelAggregatorFactory) {
 		return new MongoDBAASAggregatorFactory(mongoDBConfig, aasServerRegistry, aasAPIFactory, submodelAggregatorFactory);
 	}
 }
